Move all admin related queries into admin.ts
