--- studio/lib/queries/companyDetails.ts
@@ -1,7 +1,0 @@
-import { groq } from "next-sanity";
-
-import { companyInfoID } from "studio/schemas/documents/admin/companyInfo";
-
-export const COMPANY_INFO_QUERY = groq`*[_type == "${companyInfoID}"][0]`;
-
-export const COMPANY_LOCATIONS_QUERY = groq`*[_type == "companyLocation"]`;
